+++ crates/steel-parser/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod lexer;
-pub mod new_lexer;
 pub mod parser;
 pub mod span;
 pub mod tokens;
